% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeddo.R
\name{aeddo}
\alias{aeddo}
\title{Automated and Early Detection of Disease Outbreaks}
\usage{
aeddo(
  data,
  formula,
  k,
  sig_level,
  exclude_past_outbreaks,
  init_theta,
  lower,
  upper,
  method = "BFGS"
)
}
\arguments{
\item{method}{}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function performs automated an early detection of disease outbreaks
(aeddo) on a time series data set. It utilizes hierarchical models in an
innovative manner to infer one-step ahead random effects. In turn, these
random effects are used directly to characterize an outbreak.
}
