% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_nll_poisson_gamma_inputs}
\alias{check_nll_poisson_gamma_inputs}
\title{Check and assert the inputs for the nll_poisson_gamma function.}
\usage{
check_nll_poisson_gamma_inputs(theta, data, formula)
}
\arguments{
\item{theta}{Numeric vector of parameters.}

\item{data}{A data frame containing the time series data.}

\item{formula}{A model formula for the fixed effects in the hierarchical
model.}
}
\value{
Returns nothing. Raises errors if the inputs do not meet the
specified criteria.
}
\description{
This function is for internal use and checks the validity of inputs to the
nll_poisson_gamma function. It performs various checks using the checkmate
package to ensure that the inputs conform to expected types and values.
}
\examples{

\dontrun{
# This function is for internal use and is called within the
# nll_poisson_gamma function. It is not intended to be called directly by
# users.
}
}
