% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeddo-package.R
\docType{package}
\name{aeddo-package}
\alias{aeddo}
\alias{aeddo-package}
\title{aeddo: Automated and Early Detection of Disease Outbreaks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A powerful tool for automating the early detection of disease outbreaks in time series data. It utilizes advanced statistical method, including hiearchical models, in an innovative manner to effectively characterize outbreak signals. It is particularly useful for epidemiologists, public health professionals, and researchers seeking to identify and respond to disease outbreaks in a timely fashions. For reference on hiearchical models, see Henrik Madsen and Poul Thyregod (2011), ISBN:9781420091557.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ssi-dk.github.io/aeddo/}
}

}
\author{
\strong{Maintainer}: Kasper Schou Telkamp \email{ksst@ssi.dk} (\href{https://orcid.org/0009-0001-5126-0190}{ORCID})

Other contributors:
\itemize{
  \item Statens Serum Institut, SSI [copyright holder, funder]
}

}
\keyword{internal}
