% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeddo-package.R
\docType{package}
\name{aeddo-package}
\alias{aeddo-package}
\alias{_PACKAGE}
\title{aeddo: Automated and Early Detection of Disease Outbreaks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A powerful tool for automating the early detection of disease outbreaks in time series data. 'aeddo' employs advanced statistical methods, including hierarchical models, in an innovative manner to effectively characterize outbreak signals. It is particularly useful for epidemiologists, public health professionals, and researchers seeking to identify and respond to disease outbreaks in a timely fashion. For a detailed reference on hierarchical models, consult Henrik Madsen and Poul Thyregod's book (2011), ISBN: 9781420091557.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ssi-dk.github.io/aeddo/}
  \item \url{https://github.com/ssi-dk/aeddo}
  \item Report bugs at \url{https://github.com/ssi-dk/aeddo/issues}
}

}
\author{
\strong{Maintainer}: Lasse Engbo Christiansen \email{lsec@ssi.dk} (\href{https://orcid.org/0000-0001-5019-1931}{ORCID})

Authors:
\itemize{
  \item Kasper Schou Telkamp \email{kasper.telkamp@gmail.com} (\href{https://orcid.org/0009-0001-5126-0190}{ORCID})
}

Other contributors:
\itemize{
  \item Statens Serum Institut, SSI [copyright holder, funder]
}

}
\keyword{internal}
